Refactor factories for: NEHRP-2009, ASCE7-10, IBC-2012, IBC-2015 into NSHMP-2008 factory class
